New Unicode-capable pdButton control
I'm planning to develop specialized "OK" and "Cancel" variants (settable
via property), and of course proper animation is still on the to-do
list.  But for now, the button should integrate pretty nicely with the
default Windows 10 theme.

But more importantly: this is one more custom control down!  I'll
shortly be replacing various standard and jcButton instances with the
new control.