File > Print: fix behavior on Windows 10
I'm not actually sure, but printing might have been broken across all
OSes.  :/  I can't remember the last time I printed an image, so this is
a feature that receives very little testing!  Thanks to Frans for
raising this issue.

Still to-do is a full-featured, custom print wizard, but that's a ways
down the priority list ATM.