Complete the great command button purge
Among other things, this shaved about 100 k off the executable, so we're
improving a lot of things by moving ever closer to a completely custom
UI toolkit.  This change also allowed me to drop a bunch of .frx files,
which is always a good thing.

Many thanks to Juned Chhipa, whose "jcButton" user control served PD
well for many years, but is now being retired in favor of pdButton.