Color selection UC: overhaul UI rendering and add caption support
The color selector control was the first custom control I ever built,
and PD controls have come a long way since then.  The old version didn't
support any notion of theming or backbuffering (it still relied on
AutoRedraw!), and because I wanted to add caption support, this was a
good time to completely overhaul the control.

I still have some other ideas for this little color selector wrapper,
including a mini-color-selector on first click or possibly overlaying
the RGB or Hex value on MouseOver, but at least this is a good start
toward modernizing the control.