Effect Preview UC: lay groundwork for adjusting preview quality in real-time
One of the ongoing difficulties with PD is the variation in performance
for different effects and adjustments.  Some are fast enough to happily
render in real-time on any modern PC, while others struggle,
particularly those that involve performance-intensive operations like
supersampling.

I've got some ideas for adjusting preview quality in real-time, by
measuring preview framerate and dynamically shrinking the preview area
to match.  (This is called "dynamic resolution scaling", and I got the
idea from idTech 5 where it's used to preferentially maintain a game's
framerate over output resolution.)  This commit doesn't do that yet
(sorry), but it includes a bunch of refactoring to make such a solution
viable in the future.

That said, this commit still brings immediate benefit to preview
rendering performance, particularly when in 1:1 zoom mode.  I've done
everything I can to remove the preview user control's performance from
the drawing equation, and the result is much more responsive live
previews, particularly on low-end systems.  This was accomplished by
implementing a bunch of new optimizations for DIB creation vs reuse
(reuse is always attempted, if possible), skipping the use of WM_PAINT
and rendering immediately to the screen, implementing manual preview
backbuffering (the old solution relied on AutoRedraw!), and moving the
rendering code out of pdDIB and into the preview control itself, which
now has an optimized function specific to its purpose.