Canvas and viewport engines: overhaul in preparation for paint tools!
This involved rewriting a ton of performance-sensitive code, which had
me pretty nervous... but so far, things look good!  In fact, some of the
changes I've made should actually *improve* viewport interactions
(particularly all the math involved in converting data between image and
canvas coordinate spaces), so hip-hip-hooray for that.

The biggest user-facing change of this overhaul is that you can now
scroll past the edge of the image, which was a big prerequisite for
paint tools.  In fact, the whole way PD handles viewport/image
interactions has changed.  PD previously had like a dozen places that
calculated their own coordinate math, which was pretty damn dumb.  Now,
everything is routed through two master functions in the Drawing module,
so that should remove a lot of potential for canvas interaction bugs.

Also, the viewport is now officially subpixel-compatible from beginning
to end, which is a *huge* get.  A number of tools (like image panning)
have been updated to reflect this new capacity.  There may be some other
places in the code that still use integers where they should be using
floats, but I'll get to those soon, I hope.

The biggest ongoing problem with most viewport stuff is the outdated
selection engine.  I've got a lot of plans for fixing selections, but as
it's a massive project, I'd rather get core paint tools implemented and
push the 7.0 release out the door.  So this is more of an FYI that the
current selection code is a mess, but I'm aware and do have plans to fix
it.

In the meantime, I've got a few known bugs with the new code to fix
(like restoring the nice behavior of viewport-preservation when using
mousewheel to zoom), so I'll wait to sync this until those are in order
too.