pdCompositor: myriad performance improvements + bugfixes
GDI+ rendering issues make it very difficult to reliably solve the
"transparent checkerboard bleeds over image edges" problem, but this
update errs on the side of caution and tries to under-draw the
checkerboard if it has to, to guarantee it doesn't bleed past the edges
of solid images.

I've also reworked a ton of compositor code in light of the new viewport
engine, and performance continues to trend upward.  Some shortcuts are
now made during subpixel rendering calculations, so subpixel
measurements are cached as much as possible.  For tiny movements, the
compositor (at its default performance setting) will now avoid
recalculating subpixel data, which makes the compositor a lot faster,
but some tiny movements (e.g. scroll bars when way zoomed out) may not
perfectly represent true subpixel measurements.  I'm okay with this, as
the user can always set a preference for "max quality" viewport
rendering to get 100% accurate results.

I've commented out the on-canvas shadow around the image for the time
being, but will restore it soon.