Remove floating toolbars option
I'd like to someday reinstate this option, but as it is, I haven't given
it attention in years and it introduces a ton of buggy,
difficult-to-maintain code into the window manager.  I'm also not sure
that it even works as it should, and because the option is so hidden,
hardly anyone makes use of it.

Dumping the option will allow me to focus my efforts on making the
"docked window" experience better than it currently is, without
requiring such an insane amount of code to differentiate between
"floating" and "docked" modes.

For the next release, I'm trying hard to simplify PD's overall design,
and the window manager is an obvious place to start.