Complete window manager code purge
With floating toolboxes gone, a lot of floating-window-specific code
could be purged.  This includes a lot of complex, difficult-to-maintain
functions and interactions, so this is a nice change.

The main pdWindowManager class could probably be purged even more, but
because it would require further testing across all supported versions
of Windows, I'm not willing to mess with it right now.  Perhaps in the
future...