pdDisplays: add support for system DPI setting calculations
This is a key part of solving some long-standing DPI scaling issues with
various PD UI elements.

Also, I'm slowly working on standardizing function names as always being
capitalized (vs variable names, which are always uncapitalized), so bear
with me as the next few weeks of commits include a lot of case-changes.