pdWindowSize: new helper class for DPI-aware manipulation of user controls
I hate to resort to this, but VB's internal measurements (and
accompanying size/position functions) are hopelessly broken on any DPI
where TwipsPerPixel doesn't resolve to a whole number.  While there are
workarounds, they are clumsy and hackish, so I'd rather just rely on API
functions to handle all movement and positioning code.

To that end, user controls can now attach a pdWindowSize class to handle
movement/size data for them.  A few helper functions can also be exposed
to simplify move/size requests for outside callers, as relevant.  I
think this is the least of many evils when it comes to creating proper
high-DPI support across PD's wide range of UCs.