Improve high-DPI support for pdCanvas, pdComboBox, and pdLabel
pdComboBox is proving to be a nasty piece of work.  Given the
complexities involved in trying to synchronize a system combo box with
an underlying VB control, I'm heavily considering wiping my existing
work and writing my own combo box from scratch.

With that in mind, there are still some small pixel-level issues with
certain combo boxes that dynamically set their size according to their
contents.  Fixing this is on my to-do list.