Screen capture: minor fixes to improve Windows 8.1/10 capturing
The whole way this tool functions on Win 10 probably needs to be
revisited, but in the meantime, this commit enables capturing of
DirectX-style windows (assuming the window response to WM_PRINT
messages, which is never guaranteed).  This allows some universal
Windows apps to be captured correctly, regardless of their z-order on
the desktop.