Layer toolbox: start prepping for collapsible panel support
The right-side toolbox is going to be a lot more useful by the time 7.0
releases.  My goal is to finally implement independently collapsible
sections with some new tools PD's been lacking for too long (a Navigator
panel, for example).  This is also a necessary step to preparing for a
selection tool overhaul, where selections work more like layers (with
independently editable selection objects that can be merged via various
set operators).

To that end, the main toolbox form is now just a minimalist container
for various subforms, each of which contains a single collapsible panel.
New helper classes for sub-window management will make this much easier,
but still to-do is a new clickable label+button UC that will handle the
actual collapsing/expanding of each panel.

At present, the dialog works pretty much how it did before, except that
the layer toolbox / blend mode / opacity controls now live on an
entirely separate form that's dynamically synchronized to the toolbox's
size and shape, as necessary.