Minor bugfixes and additional hardening against potential window manager errors
There are some weird window manager issues when the image tabstrip
orientation is changed and the program is restarted, but I'm having
trouble tracking down an exact cause.  In the meantime, these fixes
prevent a hard crash.

(Fun sidenote: PD should now be runnable with the IDE set to "Break on
all errors", thanks to rewriting a bunch of code that relies on error
handling for proper behavior.)