Finally crack the math for API-accelerated rotation + edge clamping
This was unpleasant, but it opens up a large range of possibilities for
extending existing filters.

The basic idea is this: many filters operate on a horizontal+vertical
grid (e.g. blur).  These operations could be much more useful and/or
interesting if variable angles were supported, but attempting to
interpolate something like blur at any angle incurs a massive hit to
speed.

For an operation like motion blur, where angle support is a necessity,
we can optimize the filter by rotating the image once, applying an
optimized horizontal/vertical blur, then rotating the image back to its
original orientation.  The problem with this method, as usual, is how to
handle image edges.

When rotating an image into a new container, there are large, blank
corners where the rotated image doesn't fill the new container.  This
introduces blank pixels into any horizontal or vertical calculations,
which screws up all calculations because those blank pixels didn't exist
in the original image!

The highest-quality workaround is to stretch out the edges of the
original image to cover the bare areas left behind by rotation.  PD has
long provided a way to do this, but it's incredibly time-consuming, as
we have to manually interpolate each pixel of the bare regions, and VB's
SafeArrays incur a heavy penalty during interpolation since we have to
do at least four array accesses for each interpolation, then a bunch of
floating-point math to calculate accurate colors.

It occurred to me that the "edge clamp" step could be accelerated using
a parallelogram approach, and something like PlgBlt (but antialiased and
interpolated, obviously).  GDI+ offers such a function, but working out
the math for all four image edges, while also taking into account GDI+'s
propensity for things being off by subpixel amounts, wasn't pretty.

Anyway, the function is finally working, and I've already integrated it
into PD's motion blur tool (which cut the existing code amount by dozens
of lines; always nice!).

The speed increase of this new version massive; on a 6 megapixel 32-bpp
image, the latest versions of GIMP and Paint.NET clock in at 33 and 58
seconds respectively, for a 200 px motion blur.  PD does the same thing
in just over two seconds - an order of magnitude faster, and with
identical results, including proper handling of alpha bytes.

This new function allows variable angles to be extended to any grid
function with little-to-no speed penalty, which is a massive win, let
alone the speed increases it provides to existing functions like motion
blur.