Generic filters: new horizontal-only IIR implementation
This highly-optimized IIR blur can be used to blur in a single
direction, either asymmetrically (e.g. [n] pixels to the right), or
symmetrically (e.g. [n] pixels both right and left).

Directional blur is used in a number of different PD filters, and IIR
gives a much more natural result than box blur, at a small speed and
memory trade-off.  However, the memory trade-off is basically a wash, as
IIR is an in-place transform, meaning we do not need separate source and
destination images to obtain correct results.