Effects > Blur > Motion Blur: provide Gaussian (IIR) option
This greatly improves the utility of Motion Blur.  The new Gaussian
option causes the blur to "fall-off" as its distance from the pixel
increases, causing a much more natural blur.  The end result is
identical to Photoshop's methodology, and similar to the "wind" stylize
effect they provide.  (In fact, I may look at providing a "wind" stylize
option now that this is available!)