Effects > Artistic > Film Noir: rewrite from scratch with a new dialog
One more one-click filter converted to something actually useful!

As part of the overhaul, this is PD's first effect to operate entirely
in a floating-point (HDR-friendly) luminance space.  This provides a
much nicer result when working with a black-and-white image,
particularly one meant as an homage to classic 40's noir stylings.  The
long-term goal is to add HDR support to many more tools, but that's
still a ways off.