pdScrollBar: implement mousewheel support
An external wrapper is also provided, to make it easy to sync scrollbars
to any arbitrary window.  All the window has to do is relay its wheel
messages to the RelayMouseWheelEvent() function, and the scroll bar is
automatically "attached" to that window.  This will be way easier than
the previously hacky stuff we've done with custom listboxes.