Kill the canvas drop-shadow
Shadows like this are no longer in vogue, so why keep the extra code
around?  It needed to be rewritten anyway to account for subpixel
positioning, so this is a win all-around.

Similarly, the pdShadow class that rendered the canvas drop-shadow is
very old, GDI-based stuff, and it's simply not relevant with PD's new
UI.  It's gone for good, too.