pdCompositor: refactor in preparation for paint tools, layer masks
I've been meaning to refactor the compositor class for some time, so
it's nice to get this underway.

Actual blend functions are now implemented in a separate pdPixelBlender
class.  This will allow for specific, optimized functions designed
against a single blend mode, rather than the catch-all function we use
now.  I've also refactored the order in which certain compositing
actions are calculated and/or applied, which was necessary to prepare
for layer masking (as masking requires us to pre-apply alpha values
before blending pixels).

I've tried to be very careful about this, but because boundary math has
changed, OOB errors may be a possibility.  I've done a lot of testing
and will continue to do so, but please let me know if you experience any
bugs.