pdPixelBlender: prototype blend function that only uses 1D arrays
"Darken" blend mode was the guinea pig for this round.  It's difficult
to get hard numbers on what kind of a boost this approach provides, but
in the IDE we cut render time by about 20%, plus another 20-30% from the
function-wide refactoring.  This (roughly) means a best-case improvement
of 2x when compiled, vs the old code, but that number could obviously
vary.

Best of all, playing around with this approach has given me a better
idea for how to refactor blending to prepare for masks and layer styles.
All alpha handling needs to be moved to a dedicated preprocessor sub, so
we're not duplicating tons of alpha-specific code for each blend op.
That's next on the to-do list.