pdCompositor: unify various blend mode implementations
There is now just a single master implementation of all blending
functions, which was one of the main goals of this refactoring session.
Now I don't have to worry about maintaining separate blend
implementations for things like viewport rendering vs actual layer
merging, which is much, much better.

Another big part of this commit is prepping for mask and alpha
inheritance support.  By moving all alpha-specific code out of the blend
mode functions and into a dedicated alpha function, we can apply it even
to layers that use a normal blend mode (which is obviously a primary
use-case!).