Implement alpha inheritance
What's alpha inheritance?

Quick overview:
https://userbase.kde.org/Krita/Tutorial_2#Inherit_Alpha_.28alpha_.3D_transparency.29

Longer tutorial description (toward the end):
https://krita.org/item/using-the-transform-masks/

Alpha inheritance is a crucial feature when painting, because the exact
same code (literally) is how we implement a "lock alpha" feature like
Photoshop, GIMP, etc, where a brush can be disallowed from touching
existing alpha values in a layer.  As such, I wanted to get this feature
implemented now, to get an idea of how it affects the fluidity of layer
interactions.

Thanks to the recent overhaul of the pdCompositor and pdPixelBlender
classes, the performance impact of alpha inheritance looks to be
minimal.  1D array access code is used in the main loop to bring the
performance in line with a pure C-style implementation, and after some
preliminary testing, I think the performance impact will be totally
acceptable while painting.

One step closer...