Add "center image inside viewport" button to pdCanvas
Pardon the icon.  It's a WIP.

Now that PD supports scrolling past canvas edges, there should be a way
for the user to easily re-center the image within the viewport.  The
"dead space" between the horizontal and vertical canvas scrollbars seems
as good a place as any to stick the button.  (It's a tad small, but I'm
worried about confusing users if I stick it near the "fit image within
viewport" button on the status bar.  As always, suggestions welcome.)