Implement first bits of the navigator panel
A lot of optimization work is planned to improve the way various UI
elements update themselves in response to image changes, so if the
thumbnail update seems oddly delayed, that's why.

This commit is rather large because I've reworked the way the layer
panel is notified of changes.  All "image has changed" updates are now
relayed to a master function in the right-side toolbar, and it's now
responsible for relaying those notifications to any/all relevant
subpanels.