pdColorWheel: implement hue wheel rendering
Color selection on the primary canvas needs to be as accessible and
quick as humanly possible, to spare the user from ever needing to drop
into a separate color selection window.  In this regard, PD's current
color selector simply won't cut it.

Fortunately, other paint-focused software has already done a lot of the
UI research for us when it comes to color selectors.  Most paint-focused
tools provide a range of color selection UIs, which the user can switch
between easily (e.g. MyPaint:
http://mypaint.intilinux.com/?page_id=530).  I'd also like to provide a
variety of color selectors, within reason, obviously

First, I'm tackling selectors that rely on a "wheel" UI for selecting
hue.  This commit is a placeholder where the wheel is successfully
constructed rendered on-the-fly, relative to the current control's size.

Corresponding hit-detection code forthcoming.