Effects > Light and Shadow > Cross-screen: minor optimizations
During preview mode, we don't need to be so fanatical about releasing
DIBs, because they're just going to be re-used anyway.  Re-creating them
on every preview render causes a ton of memory churn.

Instead, let's postpone any forcible releasing until a non-preview
version of the effect is required, or the form is unloaded.  (Whichever
comes first.)