pdColorVariants: new user control
The end goal is to make main-screen color selection as quick, versatile,
and comprehensive as possible.  Because these three goals are
diametrically opposed, some experimentation will likely be required to
get this right.

pdColorVariants is a first stab at giving the user a quick and easy way
to "nudge" a color toward something they might like better.  The goal is
something a like mixing colors on a palette, although right now I'm
using pretty generic color models to generate the available palette
choices.  (This is a major area of research, with lots of debate over
how to do it correctly, and I'm still tracking down papers that might be
helpful.  Final algorithm TBD.)

On the top and bottom of the new control, choices are provided for +/-
hue, saturation, and luminance, while left and right provide options for
+ / - red, green, and blue.  The control is synchronized to the
pdColorWheel next to it, so you can use either control to play with
colors.

Given the speed with which painters swap between colors (e.g.
https://www.youtube.com/watch?v=93lMLEuxSLk), there can't be any delay
with these controls, so efforts have been made to keep performance
pretty much instantaneous.

Still to-do is some kind of easy-access color history, like that YouTube
link nicely demonstrates.  That panel's getting scrunched for vertical
space, however, so I likely need to implement vertical panel resizing
soon.  <sigh>