pdColorVariants: switch to a path-based system for rendering and hit-detection
This is a lot more elegant, as it allows me to use totally different
graphical layouts without any significant changes to the underlying
code.  (Only the region construction code has to change; rendering and
hit-detection does not.)

This commit also improves tooltips and various other small items.

Next up is a circular layout, so the control can mimic the color wheel
next to it.  (I hope to make switching between circular layouts and
rectangular layouts as easy as toggling a property.)