Main color selector control: re-architect in light of new color features
Lots of under-the-hood changes to improve performance and appearance,
and all color selectors now provide a right-side button that allows the
control to copy the current main window paint color.  I think this is
the best compromise between letting each control manage a unique
last-used color (which may be specific to that control's purpose), while
also making it easy to grab the primary paint color, if they need it.