pdTitle: convert to ucSupport
This is a ridiculous amount of work, but it's cutting a lot of code in
each UC, while also adding correct high-DPI support, so it's worth the
effort.  pdTitle was a good candidate for conversion since it's
relatively simple, but it paints its caption in a non-standard position,
so some new helper functions had to be written (and they'll be helpful
for future controls).

The master user control support class also handles key events now, and a
bad pdFontCollection bug was fixed.