pdLabel: convert to ucSupport
This was the largest and ugliest control conversion I have to do, mostly
because of the damn .AutoSize property that labels possess.  It's only
used in a few places throughout the program, but I needed to make sure
this overhaul didn't break that.  (Let alone other weird stuff, like
auto-fitting wordwrapped text.)

The end result is a pdLabel control that's 350 LoC lighter, and a cold
start GDI object reduction of ~150.  Labels are also DC-less now, and
given their prevalence throughout the project, that's a nice resource
reduction to have.

The main user control support class also received quite a few changes as
part of the overhaul.  It now suppresses redraw requests when a control
is invisible, which should improve all dialog load times (as controls
will now only redraw once, prior to being shown).

I also rewrote a bunch of generic caption code to make future RTL
translation support easier to handle automatically.