pdClipboard: finish overhaul of DDB, DIB, DIBv5 formats
What a disastrously complicated project this turned out to be.  The
clipboard is a hot mess when it comes to the old CF_BITMAP format, and
trying to find a solution that works without user prompts was not easy.

The biggest problem with CF_BITMAP is that the pixel data is - by
definition of a DDB - device-specific.  Modern display drivers typically
run in 32-bpp mode for performance reasons, but because the alpha bytes
are irrelevant, you'd expect them to be a nice, obvious value like 0 or
255.  Unfortunately, depending on the display driver, alpha bytes can be
all kinds of bizarre values, and short of prompting the user, you have
no idea if those values are intentional or accidental.

As such, after a lot of research and testing, PD will now enforce strict
alpha requirements on DDBs coming from the clipboard - but ONLY if the
original application actually placed a DDB there.  (If the program
placed a DIB or DIBv5 on the clipboard, we can now detect that and
preferentially take the DIB(s) over the DDB.)  Alpha is forced to 255 to
avoid driver-specific issues with the Print Screen key, so this should
finally resolve some ugly trouble with screenshots on certain OS + OS
Theme + video card + driver combinations.