FreeImage: add new function for wrapping DIBs without copying them
FreeImage added this feature in 3.17.0, and it's a great fit for PD
because it allows us to pass DIBs to FreeImage without doubling our
memory requirements (or wasting time allocating duplicate chunks of
image-sized memory).

I've only added this new feature to Clipboard > Copy for the moment,
when PD creates PNG clipboard data for interop, but I'll soon be
replacing other FreeImage interfaces with the new code.  This should
speed up quite a few things throughout the program.