Implemented delayed rendering for clipboard objects
Originally, I just wanted to rewrite PD's clipboard class because I was
using a third-party class from a website known to produce buggy code.
But one thing led to another, and before I knew it I was reworking the
whole way PD manages its clipboard interactions.  While I'm "in the
zone" with clipboard stuff, I may as well go all the way, which leads us
to this commit.

Delayed rendering is a cool clipboard trick where you don't actually
place data on the clipboard until a program actually requests the data.
For example: PD supports multiple image formats on the clipboard.  If a
user is editing a 10-megapixel image and they hit "copy", PD needs to
stick a bunch of 10-megapixel image copies on the clipboard - one copy
for each supported image format.  This is a lot of work, and it might
take a few seconds.  Worse still, the clipboard now takes up a ton of
resources, because it's holding a bunch of massive image copies in
various formats.

With delayed rendering, when the user clicks "copy", PD simply tells the
clipboard "hey - if a program wants PNG, bitmap, DIB, or DIBv5 data, let
us know."  In addition, PD "stashes" a copy of the current DIB (with
selections and such processed) to file, in case the on-screen image
changes between now and when someone hits "Paste".

Then PD waits.

If the user switches to another program and hits "Paste", Windows comes
running back to PD, screaming, "hey - a program wants to paste PD's
data.  Here's the format they requested."  PD now intercepts that
message and generates a copy of the requested data on-the-fly, and the
other program ends up with exactly what it wants.  The clipboard is
never required to hold more formats than those specifically requested
since the last "Copy" action, and to the user, this is all handled
seamlessly.

Delayed rendering is a great mechanic, but it's somewhat cumbersome in
VB since we can't easily trap window messages like WM_DESTROY.
WM_DESTROY is important because if the user shuts down PD while PD is
still responsible for clipboard data, PD needs to formally upload all
supported clipboard data before exiting.  (Otherwise, if the user hits
"Paste" in some other program, PD won't be around to intercept their
request!)  So some manual work is required to cover this case, while
also doing little things like deleting PD's local clipboard stash if
it's no longer required, reworking "Copy" and "Cut" to pass delayed
messages, and setting up event handlers to respond to WM_RENDERFORMAT
messages.

Anyway, the take-home message is that "Copy" and "Cut" actions are now
extremely fast, especially if the user is pasting the data into another
program.  If they're pasting the data back into PD, performance is still
mediocre, because when "Paste" is hit, we render a PNG on the fly, then
decompress it on the fly, which is asinine.  I'll shortly be adding a
custom PD clipboard format that bypasses all this, making Copy+Paste
interactions within PD much much faster.

As part of this work, PD's shutdown sequence was reworked a bit; that's
why some other random files are included.