Provisional PSD export support
LAYERS ARE NOT CURRENTLY SUPPORTED.  This is a FreeImage limitation, and
I've got too much on my plate to add layer functionality to their DLL.
I've started some discussions in their support forums regarding layer
support (since they already support multipage documents, which provides
a viable API), and though work has been done on this in the past and it
was very close to getting merged, I haven't gotten any feedback on why
they ultimately abandoned it.  :(

That said, basic single-layer PSD export is now supported.  I've
manually compiled FreeImage to enable this new support, so bugs in the
DLL are possible because this is a development build based off
development code.

The few images I've tested open okay in PD itself, GIMP, and Irfanview.
Additional testing is of course welcome.