pdPixelIterator: initial build
Many PhotoDemon effects and filters require us to search some radius (r)
around each pixel.  Blur is an easy example - to blur a pixel, you visit
each pixel within the blur radius, add together all their values, and
set the center pixel to the average value of the region.

By default, area filters have O((2r+1)^2) complexity, because you must
search through (2r+1) * (2r+1) pixels for each source pixel in the
image.  This incurs a massive speed penalty, so a lot of work in PD goes
into improving the performance of these filters.

The primary optimization PD uses is a serpentine "sliding window"
approach.  This technique is illustrated nicely on page 3 of this PDF
(http://www.shellandslate.com/download/fastmedian_5506.pdf).  This
technique reduces complexity of the algorithm to roughly O(r), by
reusing the results of past calculations, and performing calculations
using pixel-agnostic histograms.

Besides the speed boost, I love this technique because it is easily
generalized, which allows me to apply it to many different PD functions
with only minimal changes.  (At present, more than a dozen PD effects
and filters use a copy+paste version of this code.)

With that in mind, the purpose of this new pdPixelIterator class is
twofold.  First, the class takes the generalized serpentine sliding
window code, and abstracts it into a standalone class.  This allows us
to cut some 200 lines of code from every function that uses the approach
- always great, since that was a lot of copy+paste messiness!

This part of the class is complete.  Some weird array tricks are used to
avoid the need for CopyMemory, so the class writes its histogram data
directly into histogram arrays supplied by the caller.  This should make
performance roughly identical to the old copy+paste method, despite the
new abstraction.  (And future optimizations may actually give us a speed
bump over the old method; I'm holding off on this until I deal with part
2, described below.)

The second purpose of the class is to implement new functionality.  The
serpentine sliding window approach is also great because it is not
limited to square-shaped regions.  Any isotoxal polygon (squares,
circles, stars, triangles, etc) at any angle is a valid input.  As an
added bonus, different shapes do not affect the performance of the
algorithm, and in fact they may be faster than square regions if they
cover fewer pixels in each region.

I have not yet implemented this "shape" feature, but pdPixelIterator is
already designed against the possibility of supporting any general
shapes.  This feature will allow for many new possibilities, ranging
from the obvious (circular pixel regions instead of square ones, for a
much more natural end result) to the artistic (bokeh of any shape,
woohoo!).  By implementing this shape functionality into a standalone
class, any function that uses pdPixelIterator will be able to support
ALL shape options with no work on the part of the caller.

Shape functionality is next on my to-do list.

The test function for this initial implementation is Effects > Noise >
Mean-Shift, because it was added during this release cycle, it's
computationally expensive, and it's easy to tell if it's working or not.
So far, all looks good.  Pending additional testing, I'll start rolling
pdPixelIterator out to other filters in the coming weeks.