Median filter: convert to pdPixelIterator
This is a significant improvement, as the core median function is used
by a bunch of other filters (erode, dilate, trace contour), as well as
the raster-based Modify Selection functions (grow, shrink, etc).  The
selection functions in particular produce a much, much better result
when a circular kernel is used, so I'm quite happy to have this
implemented.

(Also, circular median kernels was one of Paint.NET's few advantages
over PD - always nice to catch up!)