Effects > Noise > Anisotropic Diffusion: new tool
For a nice gallery of anisotropic filtering uses, check out:
http://www.cs.utah.edu/~manasi/coursework/cs7960/p2/project2.html

As noted in the comments, this is basically a reference implementation
of Perona and Malik's original paper
(http://authors.library.caltech.edu/6498/1/PERieeetpami90.pdf).  It's a
pretty fast implementation despite an insane amount of array accesses,
and it supports a variety of inputs, which I've tried to rename and/or
redistribute to make the tool more approachable for beginners.

Anisotropic diffusion works by calculating per-pixel energy gradients in
multiple directions (PD supports both cardinal and ordinal), then
merging neighboring pixels based on the "slope" of their color
difference.  If high slopes are favored, you get a very nice sharpening
filter that operates only on high-contrast edges; if low slopes are
favored, you get high-quality noise-removal that safely ignores edges.

Because anisotropic filtering exists in a weird place (as both a blur
filter and a sharpen filter), I've stuck it in the Noise menu for now,
as its smoothing result bears some similarity to filters like Mean-Shift
or Bilateral Smoothing, while its sharpening is somewhat like an "add
noise" filter that's restricted to edges of a certain contrast
magnitude.