Clipboard interface: fix for CF_BITMAP if PD is closed prior to Paste operation
This is explained in more detail in the code, but basically, if PD is
shut down while it's still the owner of the clipboard's data, we have to
manually reverse scanline order for the CF_BITMAP format.  This doesn't
need to be done if PD is still running when another program initiates a
Paste - only if the order is specifically Copy-Inside-PD, Close PD,
Paste into some other program.