Adjustments > Photo > Red-eye removal: implement first stage false-positive removal
This false-positive removal step works by comparing each potential
red-eye region to the pixels surrounding it.  If a significant degree of
pixel similarity is found (e.g. > ~10%), we discard the entire region,
as we've probably found a skin or scenery element that's part of a
larger pattern (vs a standalone eye or eye pair).

This is arguably the most important false-positive removal stage, as
it's...

1) Not prone to false-negatives
2) Easy to implement a UI toggle for the method's sensitivity (as a
failsafe only; the default value actually works very well)

On my set of test images, this test removes 80-90% of false-positive
regions.

Next up are additional checks for invalid region shapes and sizes, which
are much easier to detect and remedy.