Adjustments > Photo > Red-eye: finish core correction framework
This commit includes the dynamic region merge code that combines
highlight and red-eye regions into contiguous "eye" shapes, as well as a
basic implementation of the original paper's suggested correction
algorithm (which I'm not thrilled with, but I'm still investigating
replacements).

There's still work to be done to make the result look "natural" (e.g.
blending along rough edges), but as of this commit, the most troublesome
portions of the algorithms are complete.