Adjustments > Photo > Red-eye: add antialiasing, performance improvements
There's more that could be done to improve this filter, but the code's
become somewhat of a rat's nest, and integrating the tool into PD has
taken way longer than I intended.  So I'm going to move on for now, with
plans to revisit when things have quieted down a bit.