pdLabel: finalize theming support
With this commit, I can now start serious theme testing via the Tools >
Developers > Theme Editor menu.  The dark theme at this point is
(literally) a cut-and-paste version of the light theme, with dark and
light shades reversed, but I'll start editing the file seriously once I
have a few more controls to play with.

In the meantime, the few controls I've finished at least confirm that
the theme engine is correctly loading and parsing the theme XML files,
and the IDE can still be used without everything exploding...