Fix control name inconsistencies
As part of this long-coming fix, there's a new (quick-and-dirty) tool in
the /Support subfolder for applying universal search-and-replace inside
a classic VB project.  Multiline search-and-replace is supported
(although tabs are not treated specially, so you'd need to manually
match indentation), but more relevant for me is replacing text inside
Class/Control/Form/Module headers.  This is incredibly helpful when
doing something like renaming user controls, as their instance names
can't be replaced from inside the IDE (unless you do it manually for
each occurrence, ugh).